requirements.txt
fastapi==0.115.12
google-genai==1.12.1
langchain==0.3.24
langchain_google_genai==2.1.3
neo4j==5.28.1
nltk==3.9.1
peft==0.14.0
pymupdf==1.25.5
python-dotenv==1.1.0
requests==2.32.3
spacy==3.8.5
streamlit==1.44.1
torch==2.5.1
transformers==4.51.1
uvicorn==0.34.2

# spacy.cli.download("en_core_web_sm")
# python -c "import nltk; nltk.download('punkt'); nltk.download('punkt_tab');"